Correct  rotate log

+++ main/log_store.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-void log_add(uint8_t event_id, int port_id, uint64_t value, int64_t ts);
+void log_add(uint8_t event_id, uint8_t port_id, uint64_t value, int64_t ts);
 //uint8_t event_id, uint8_t port_id, uint64_t value
 //void log_add_door_event(int door_id, int is_open);
 //void log_add_rex_event(int door_id);
